Changes to the main.py & calculations.py
On main.py
Added function to verify if the input requested is a float, if not, returns a message asking for a correct value.
Added functions to verify if the input requested is a valid option, if not, returns a message asking for a correct option.

On calculations.py
Deleted unused comments.
Created function to repeat the request of a value. If the value is a float returns the value, if the value is not a float, returns a message to ask for a correct value.
Created function to repeat a question. If the input is correct returns the option, if not, ask for a correct option.
Added float question on both cases of quantity pay in dollars.
Added else statement on the main function of request for retention and payment on dollars.
Added todo list with bills formats.
Corrected the percentage of the IVA payed.